Rebase enumerations on integer datatypes - Int is the base type for enumerations - UInt is the base type for bitfields This matches the return types of the GObject.Value.GetEnum / GetFlags methods.
With GTK3 support this was problematic as there were some enumerations (int based) which defined values in the range of unsigned integers. This was the reason long was used as it could handle all values which could be created either with integer or unsigned integer

Starting with GTK4 such enumerations are not longer used and the base type can be switched to int / uint.

In case such enumerations appear again those enumerations would probably need to be dynamically declared as uint based instead of int based. During runtime this must then be evaluated and handled accordingly (e.g. via Unsafe.As<int, uint>()).

+++ src/Generation/Generator/Renderer/Public/Bitfield.cs
@@ -21,7 +21,7 @@ namespace {Namespace.GetPublicName(bitfield.Namespace)};
 
 {PlatformSupportAttribute.Render(bitfield as GirModel.PlatformDependent)}
 [Flags]
-public enum {bitfield.Name} : ulong
+public enum {bitfield.Name} : uint
 {{
     {bitfield
         .Members
